# backend/docs/swagger.yaml
basePath: /api
definitions:
  model.AddFriendReq:
    properties:
      receiver_id:
        example: "0"
        type: string
      sender_name:
        maxLength: 64
        type: string
      verification_message:
        maxLength: 128
        type: string
    required:
    - receiver_id
    - sender_name
    type: object
...
  model.ConversationListResp:
    properties:
      content:
        type: string
      conversation_id:
        example: "0"
        type: string
      remark:
        type: string
      unread_count:
        type: integer
    type: object
...
  model.FriendRequestListResp:
    properties:
      created_at:
        type: string
      request_id:
        example: "0"
        type: string
      sender_id:
        example: "0"
        type: string
      sender_name:
        type: string
      status:
        type: string
      verification_message:
        type: string
    type: object
  model.FriendshipListResp:
    properties:
      friend_id:
        example: "0"
        type: string
      friend_remark:
        type: string
      friendship_id:
        description: friend_ships 表的主键
        example: "0"
        type: string
    type: object
...
  model.LoginByPhoneReq:
    properties:
      password:
        maxLength: 72
        minLength: 6
        type: string
      phone_number:
        type: string
    required:
    - password
    - phone_number
    type: object
  model.LoginByUidReq:
    properties:
      password:
        maxLength: 72
        minLength: 6
        type: string
      uid:
        maxLength: 20
        minLength: 1
        type: string
    required:
    - password
    - uid
    type: object
  model.LoginResp:
    properties:
      token_class:
        $ref: '#/definitions/model.TokenResp'
      user_info:
        $ref: '#/definitions/model.UserInfoResp'
    type: object
...
  model.RegisterReq:
    properties:
      name:
        maxLength: 64
        minLength: 1
        type: string
      password:
        maxLength: 72
        minLength: 6
        type: string
      phone_number:
        type: string
    required:
    - name
    - password
    - phone_number
    type: object
...
  model.TokenResp:
    properties:
      expires_in:
        type: integer
      refresh_token:
        type: string
      token:
        type: string
    type: object
...
  model.UserInfoResp:
    properties:
      name:
        type: string
      uid:
        type: string
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API documentation for the Inquire backend.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Inquire API
  version: "1.0"
paths:
  /auth/conversations:
    get:
      consumes:
      - application/json
      description: Get the list of active conversations
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            items:
              $ref: '#/definitions/model.ConversationListResp'
            type: array
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取会话列表
      tags:
      - conversation
...
  /auth/friendship_requests:
    get:
      consumes:
      - application/json
      description: 获取当前收到的好友申请列表
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            items:
              $ref: '#/definitions/model.FriendRequestListResp'
            type: array
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取好友申请列表
      tags:
      - friendship_request
    post:
      consumes:
      - application/json
      description: 发送好友申请给指定用户
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 添加好友请求体
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.AddFriendReq'
      produces:
      - application/json
      responses:
        "201":
          description: 发送成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 发送失败
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: 请勿重复发送
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 发送好友申请
      tags:
      - friendship_request
...
  /auth/friendships:
    get:
      consumes:
      - application/json
      description: 获取当前用户的好友列表
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            items:
              $ref: '#/definitions/model.FriendshipListResp'
            type: array
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取好友列表
      tags:
      - friendship
...
  /auth/refresh_token:
    post:
      consumes:
      - application/json
      description: 使用 Refresh Token 获取新的 Access Token
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: 刷新成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.TokenResp'
              type: object
        "500":
          description: Token问题
          schema:
            $ref: '#/definitions/response.Response'
      summary: 刷新Token
      tags:
      - auth
  /login/phone_number:
    post:
      consumes:
      - application/json
      description: 使用手机号和密码登录
      parameters:
      - description: 手机号登录请求体
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.LoginByPhoneReq'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.LoginResp'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 手机号登录
      tags:
      - user
  /login/uid:
    post:
      consumes:
      - application/json
      description: 使用微信号和密码登录
      parameters:
      - description: 微信号登录请求体
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.LoginByUidReq'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.LoginResp'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 微信号登录
      tags:
      - user
  /register:
    post:
      consumes:
      - application/json
      description: 用户注册接口
      parameters:
      - description: 注册请求体
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.RegisterReq'
      produces:
      - application/json
      responses:
        "201":
          description: 注册成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 请求参数错误或手机号重复
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 用户注册
      tags:
      - user
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"